hack to get around learning relational division
The title of the Kata suggests I solve this, which makes sense considering in relational algebra the solution is obviously expressible that way. However, I can't figure out how to implement relational division in SQL so I did this. I'll be able to see the solutions now that I've solved it though.